--- src/features/vocabulary/components/WordInfo/index.ts
@@ -1,1 +1,0 @@
-export { default as WordInfo } from './WordInfo'
