--- src/features/vocabulary/components/index.ts
@@ -1,2 +1,0 @@
-export { Vocabulary } from './Vocabulary'
-export { Layout } from './Layout'
