--- src/features/vocabulary/index.ts
@@ -1,1 +1,0 @@
-export { default as VocabularyPage } from './Page'
