+++ src/hooks/index.ts
@@ -1,3 +1,3 @@
 export { useActions } from './useActions'
 export { useTypedSelector } from './useTypedSelector'
-export { useFollowSeverStatus } from './useFollowSeverStatus'
+export { useFollowServerStatus } from './useFollowServerStatus'
