+++ src/pages/vocabulary/components/AddTranslationForm/AddTranslationForm.css
