+++ src/pages/vocabulary/components/Layout/index.ts
