+++ src/pages/vocabulary/components/Meaning/Meaning.css
