+++ src/pages/vocabulary/components/Meaning/index.ts
