+++ src/pages/vocabulary/components/MeaningList/MeaningList.css
