+++ src/pages/vocabulary/components/MeaningList/MeaningList.tsx
