+++ src/pages/vocabulary/components/MeaningList/index.ts
