+++ src/pages/vocabulary/components/Page/Page.tsx
@@ -1,5 +1,7 @@
 import React from 'react'
-import { Vocabulary, Layout } from './components'
+import { Vocabulary } from '../Vocabulary'
+import { Layout } from '../Layout'
+
 
 const Page = () => {
   return (
