+++ src/pages/vocabulary/components/Page/index.ts
@@ -1,0 +1,1 @@
+export { default as Page } from './Page'
