+++ src/pages/vocabulary/components/Translation/Translation.css
