+++ src/pages/vocabulary/components/Translation/Translation.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react'
 import { ITranslation } from '../../../../models/ITranslation'
 import { Manageable } from '../../../../ui'
 import { useRemoveTranslationMutation } from '../../../../store/api'
-import { useFollowSeverStatus } from '../../../../hooks'
+import { useFollowServerStatus } from '../../../../hooks'
 import './Translation.css'
 
 interface Props {
@@ -22,7 +22,7 @@ const Translation: FC<Props> = ({
     }
   ] = useRemoveTranslationMutation()
 
-  useFollowSeverStatus({ status: translationDeletingStatus })
+  useFollowServerStatus({ status: translationDeletingStatus })
 
   const handleRemoveTranslation = () => {
     removeTranslation({ meaningId: meaningId, translationId: translation._id })
