+++ src/pages/vocabulary/components/Translation/index.ts
