+++ src/pages/vocabulary/components/TranslationList/TranslationList.css
