+++ src/pages/vocabulary/components/TranslationList/TranslationList.tsx
