+++ src/pages/vocabulary/components/TranslationList/index.ts
