+++ src/pages/vocabulary/components/Vocabulary/Vocabulary.css
