+++ src/pages/vocabulary/components/Vocabulary/index.ts
