+++ src/pages/vocabulary/components/WordCard/WordCard.css
