+++ src/pages/vocabulary/components/WordCard/index.ts
