+++ src/pages/vocabulary/components/WordCardList/WordCardList.css
