+++ src/pages/vocabulary/components/WordCardList/index.ts
