+++ src/pages/vocabulary/components/WordCardsPanel/WordCardsPanel.css
