+++ src/pages/vocabulary/components/WordCardsPanel/index.ts
