+++ src/pages/vocabulary/components/WordInfoPanel/WordInfoPanel.css
