+++ src/pages/vocabulary/components/WordInfoPanel/index.ts
