+++ src/pages/vocabulary/index.ts
@@ -1,0 +1,1 @@
+export { Page as VocabularyPage } from './components/Page'
